requirements.txt
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
pgvector==0.2.2
openai==1.3.3
sentence-transformers==2.2.2
numpy==1.24.3
fastapi==0.103.2
uvicorn==0.23.2
streamlit==1.28.0
python-dotenv==1.0.0
datasets==2.14.5
pandas==2.0.3
requests==2.31.0
